# specs/001-school-library-system/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: BCD Library Management System API
  description: |
    REST API for school library management (Bibliothèque Centre Documentaire).

    **Operational Model**: Librarian-operated system with offline-first design.
    - No fines/fees (privilege blocking instead)
    - No email/SMS notifications (printed reports)
    - Librarian-mediated access (no student self-service)

    **Key Features**:
    - Circulation (checkout/return/renew with barcode scanning)
    - Cataloging (BNF SRU API integration for ISBN lookup)
    - Borrower management (students/teachers/staff organized by class)
    - Holds/reservations (librarian-mediated queue management)
    - Reports (overdue, statistics, never-borrowed)
    - Barcode generation (Code 39/128)
  version: 1.0.0
  contact:
    name: BCD API Support
    email: support@bcd-library.local
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server (localhost)
  - url: http://library-server.local:8000/api/v1
    description: Production server (local network)
...
paths:
...
  /circulation/checkout:
    post:
      tags: [circulation]
      summary: Check out items to a borrower
      description: |
        Creates circulation transactions for one or more items checked out to a borrower.
        Validates borrower is active and not over loan limit.
      operationId: checkoutItems
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - borrower_id
                - item_ids
              properties:
                borrower_id:
                  type: string
                  description: Borrower's unique ID
                  example: "101"
                item_ids:
                  type: array
                  items:
                    type: string
                  description: Array of item IDs to check out
                  example: ["785", "787"]
                  minItems: 1
                  maxItems: 10
            example:
              borrower_id: "101"
              item_ids: ["785", "787"]
      responses:
        '201':
          description: Items successfully checked out
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction_ids:
                    type: array
                    items:
                      type: integer
                    description: IDs of created circulation transactions
                  due_dates:
                    type: array
                    items:
                      type: string
                      format: date
                    description: Due dates for each item (ISO 8601)
                  borrower:
                    $ref: '#/components/schemas/BorrowerSummary'
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CheckedOutItem'
              example:
                transaction_ids: [1, 2]
                due_dates: ["2026-02-13", "2026-02-13"]
                borrower:
                  borrower_id: "101"
                  full_name: "Amira BENALI"
                  class_name: "CP-A"
                  current_loans: 2
                  loan_limit: 2
                items:
                  - item_id: "785"
                    title: "Ils ont arrêté mon père"
                    due_date: "2026-02-13"
                  - item_id: "787"
                    title: "Stuart Little"
                    due_date: "2026-02-13"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict (item already on loan, borrower over limit, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "checkout_conflict"
                message: "Borrower has reached maximum loan limit (2/2)"
                details:
                  borrower_id: "101"
                  current_loans: 2
                  loan_limit: 2
  /circulation/return:
    post:
      tags: [circulation]
      summary: Return items
      description: |
        Processes return of one or more items. Updates circulation transaction
        with return date and resets item status to available.
      operationId: returnItems
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - item_ids
              properties:
                item_ids:
                  type: array
                  items:
                    type: string
                  description: Array of item IDs to return
                  example: ["785", "787", "790"]
                  minItems: 1
                  maxItems: 20
            example:
              item_ids: ["785", "787"]
      responses:
        '200':
          description: Items successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  returned:
                    type: array
                    items:
                      type: object
                      properties:
                        item_id:
                          type: string
                        return_date:
                          type: string
                          format: date-time
                        was_overdue:
                          type: boolean
                        days_overdue:
                          type: integer
                        title:
                          type: string
                        borrower_name:
                          type: string
              example:
                returned:
                  - item_id: "785"
                    return_date: "2026-02-10T14:30:00Z"
                    was_overdue: false
                    days_overdue: 0
                    title: "Ils ont arrêté mon père"
                    borrower_name: "Amira BENALI"
                  - item_id: "787"
                    return_date: "2026-02-10T14:30:00Z"
                    was_overdue: true
                    days_overdue: 3
                    title: "Stuart Little"
                    borrower_name: "Amira BENALI"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /circulation/renew:
    post:
      tags: [circulation]
      summary: Renew items
      description: |
        Renews selected items for borrower. Extends due date by loan duration.
        Items cannot be renewed if they have holds or exceeded renewal limit.
      operationId: renewItems
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - item_ids
              properties:
                item_ids:
                  type: array
                  items:
                    type: string
                  description: Array of item IDs to renew
                  example: ["785", "787"]
                  minItems: 1
            example:
              item_ids: ["785", "787"]
      responses:
        '200':
          description: Renewal results (may be partial success)
          content:
            application/json:
              schema:
                type: object
                properties:
                  renewed:
                    type: array
                    items:
                      type: object
                      properties:
                        item_id:
                          type: string
                        new_due_date:
                          type: string
                          format: date
                        renewal_count:
                          type: integer
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        item_id:
                          type: string
                        reason:
                          type: string
                          enum: [renewal_limit_exceeded, item_has_holds, item_not_on_loan]
                        message:
                          type: string
              example:
                renewed:
                  - item_id: "785"
                    new_due_date: "2026-02-27"
                    renewal_count: 1
                errors:
                  - item_id: "787"
                    reason: "item_has_holds"
                    message: "Cannot renew - item has pending holds"
        '400':
          $ref: '#/components/responses/BadRequest'
  /circulation/borrower/{borrower_id}/items:
    get:
      tags: [circulation]
      summary: Get borrower's current loans
      description: Returns all items currently on loan to a borrower with renewal eligibility
      operationId: getBorrowerLoans
      parameters:
        - name: borrower_id
          in: path
          required: true
          schema:
            type: string
          example: "101"
      responses:
        '200':
          description: Borrower's current loans
          content:
            application/json:
              schema:
                type: object
                properties:
                  borrower:
                    $ref: '#/components/schemas/BorrowerSummary'
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        item_id:
                          type: string
                        title:
                          type: string
                        checkout_date:
                          type: string
                          format: date-time
                        due_date:
                          type: string
                          format: date
                        days_overdue:
                          type: integer
                        is_overdue:
                          type: boolean
                        can_renew:
                          type: boolean
                        renewal_count:
                          type: integer
                        renewal_limit:
                          type: integer
              example:
                borrower:
                  borrower_id: "101"
                  full_name: "Amira BENALI"
                  class_name: "CP-A"
                  current_loans: 2
                  loan_limit: 2
                items:
                  - item_id: "785"
                    title: "Ils ont arrêté mon père"
                    checkout_date: "2026-01-30T10:15:00Z"
                    due_date: "2026-02-13"
                    days_overdue: 0
                    is_overdue: false
                    can_renew: true
                    renewal_count: 0
                    renewal_limit: 2
        '404':
          $ref: '#/components/responses/NotFound'
  /circulation/item/{item_id}/history:
    get:
      tags: [circulation]
      summary: Get item circulation history
      description: Returns current borrower and past circulation history for an item
      operationId: getItemHistory
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
          example: "785"
      responses:
        '200':
          description: Item circulation history
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    type: object
                    properties:
                      item_id:
                        type: string
                      title:
                        type: string
                      status:
                        type: string
                  current_borrower:
                    type: object
                    nullable: true
                    properties:
                      borrower_id:
                        type: string
                      full_name:
                        type: string
                      checkout_date:
                        type: string
                        format: date-time
                      due_date:
                        type: string
                        format: date
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        borrower_id:
                          type: string
                        borrower_name:
                          type: string
                        checkout_date:
                          type: string
                          format: date-time
                        return_date:
                          type: string
                          format: date-time
                        days_overdue:
                          type: integer
        '404':
          $ref: '#/components/responses/NotFound'
  /circulation/borrower/{borrower_id}/history:
    get:
      tags: [circulation]
      summary: Get borrower circulation history
      description: Returns current loans and past circulation history for a borrower
      operationId: getBorrowerHistory
      parameters:
        - name: borrower_id
          in: path
          required: true
          schema:
            type: string
          example: "101"
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: Borrower circulation history
          content:
            application/json:
              schema:
                type: object
                properties:
                  borrower:
                    $ref: '#/components/schemas/BorrowerSummary'
                  current_loans:
                    type: array
                    items:
                      type: object
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        item_id:
                          type: string
                        title:
                          type: string
                        checkout_date:
                          type: string
                          format: date-time
                        return_date:
                          type: string
                          format: date-time
                        days_overdue:
                          type: integer
        '404':
          $ref: '#/components/responses/NotFound'

  # ========================================
  # HOLD/RESERVATION ENDPOINTS
  # ========================================
...
components:
  schemas:
...
    BorrowerSummary:
      type: object
      properties:
        borrower_id:
          type: string
        full_name:
          type: string
        class_name:
          type: string
          nullable: true
        current_loans:
          type: integer
        loan_limit:
          type: integer
        active:
          type: boolean
        blocked_reason:
          type: string
          nullable: true
...
    CheckedOutItem:
      type: object
      properties:
        item_id:
          type: string
        title:
          type: string
        due_date:
          type: string
          format: date
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Bad request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "validation_error"
            message: "Invalid request parameters"
            details:
              field: "borrower_id"
              issue: "Required field missing"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Resource not found"